# src/main/resources/docs/dgbank-api-docs.yaml
openapi: 3.0.3
info:
  title: DG Bank API
  description: |
    API completa per il sistema bancario DG Bank.
    
    ## Autenticazione
    L'API utilizza JWT Bearer token per l'autenticazione. Dopo il login o la registrazione, includi il token nell'header Authorization:
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Token JWT: 24 ore di validità
    - Refresh token: 7 giorni di validità
    
    ## Formati Monetari
    Tutti gli importi sono gestiti con precisione decimale (15,2) per evitare errori di arrotondamento.
  version: 1.0.0
  contact:
    name: DG Bank Development Team
    email: dev@dgbank.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.dgbank.com
    description: Production server
...
paths:
...
  /accounts/create:
    post:
      tags:
        - Accounts
      summary: Crea nuovo conto
      description: |
        Crea un nuovo conto bancario per un utente esistente. 
        Genera automaticamente un IBAN italiano nel formato: IT XX X05472 81110 XXXXXXXXXXXX
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Conto creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Errore nella creazione del conto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Utente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{accountId}:
    get:
      tags:
        - Accounts
      summary: Dettagli conto
      description: Ottieni i dettagli completi di un conto specifico
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Dettagli del conto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Conto non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /accounts/{accountId}/balance:
    get:
      tags:
        - Accounts
      summary: Saldo conto
      description: Ottieni il saldo corrente di un conto
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Saldo del conto
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  accountId:
                    type: integer
                    format: int64
                    example: 1
                  balance:
                    type: string
                    description: Importo con precisione decimale
                    example: "1250.75"
        '404':
          description: Conto non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{accountId}/freeze:
    put:
      tags:
        - Accounts
      summary: Congela conto
      description: Congela un conto bancario (imposta isActive = false)
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Conto congelato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Errore nel congelamento del conto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{accountId}/unfreeze:
    put:
      tags:
        - Accounts
      summary: Scongela conto
      description: Scongela un conto bancario (imposta isActive = true)
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Conto scongelato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Errore nello scongelamento del conto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/user/{userId}:
    get:
      tags:
        - Accounts
      summary: Conti utente
      description: Ottieni tutti i conti di un utente con il saldo totale
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista dei conti dell'utente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accounts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
                  totalBalance:
                    type: string
                    description: Saldo totale di tutti i conti
                    example: "5420.30"
        '500':
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Card endpoints
...
  /cards/account/{accountId}:
    get:
      tags:
        - Cards
      summary: Carte del conto
      description: Ottieni tutte le carte collegate ad un conto specifico
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista delle carte del conto
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cards:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
...
  /transactions/account/{accountId}:
    get:
      tags:
        - Transactions
      summary: Transazioni del conto
      description: Ottieni tutte le transazioni di un conto (sia in entrata che in uscita)
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista delle transazioni
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accountId:
                    type: integer
                    format: int64
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  count:
                    type: integer
                    description: Numero totale di transazioni
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT ottenuto tramite login o registrazione.
        Durata: 24 ore per access token, 7 giorni per refresh token.
  schemas:
...
    CreateAccountRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: integer
          format: int64
          description: ID dell'utente per cui creare il conto
...
    AccountResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Conto creato con successo"
        account:
          $ref: '#/components/schemas/Account'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operazione completata con successo"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Messaggio di errore"

    # Entity models
...
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID univoco del conto
        accountNumber:
          type: string
          maxLength: 34
          description: Numero IBAN del conto
          example: "IT 60 X05472 81110 000000123456"
        balance:
          type: string
          description: Saldo del conto (precisione decimale)
          example: "1250.75"
        isActive:
          type: boolean
          description: Stato del conto (true = attivo, false = congelato)
        dateOpened:
          type: string
          format: date-time
          description: Data di apertura del conto
        owner:
          type: object
          description: Informazioni del proprietario del conto
          properties:
            id:
              type: integer
              format: int64
            fullName:
              type: string
              example: "Mario Rossi"
            email:
              type: string
              format: email
    Card:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID univoco della carta
        cardNumber:
          type: string
          description: Numero carta mascherato per sicurezza
          example: "**** **** **** 3456"
        expiryDate:
          type: string
          format: date
          description: Data di scadenza della carta
        isActive:
          type: boolean
          description: Stato della carta (true = attiva, false = bloccata)
        owner:
          type: string
          description: Nome del proprietario della carta
          example: "Mario Rossi"
        accountNumber:
          type: string
          description: Numero IBAN del conto collegato
          example: "IT 60 X05472 81110 000000123456"
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID univoco della transazione
        transactionId:
          type: string
          description: ID pubblico della transazione (UUID)
          example: "123e4567-e89b-12d3-a456-426614174000"
        amount:
          type: string
          description: Importo della transazione
          example: "150.00"
        description:
          type: string
          maxLength: 255
          description: Descrizione della transazione
        type:
          $ref: '#/components/schemas/TransactionType'
        status:
          $ref: '#/components/schemas/TransactionStatus'
        timestamp:
          type: string
          format: date-time
          description: Data e ora della transazione
        fromAccount:
          type: object
          description: Informazioni del conto di origine
          nullable: true
          properties:
            id:
              type: integer
              format: int64
            accountNumber:
              type: string
            ownerName:
              type: string
        toAccount:
          type: object
          description: Informazioni del conto di destinazione
          nullable: true
          properties:
            id:
              type: integer
              format: int64
            accountNumber:
              type: string
            ownerName:
              type: string

    # Enums
...
    TransactionType:
      type: string
      enum:
        - TRANSFER
        - DEPOSIT
        - WITHDRAWAL
        - PAYMENT
      description: |
        Tipologie di transazione:
        * `TRANSFER` - Trasferimento tra conti interni
        * `DEPOSIT` - Deposito (da esterno)
        * `WITHDRAWAL` - Prelievo (verso esterno)
        * `PAYMENT` - Pagamento
    TransactionStatus:
      type: string
      enum:
        - PENDING
        - COMPLETED
        - FAILED
        - CANCELLED
      description: |
        Stati della transazione:
        * `PENDING` - In sospeso
        * `COMPLETED` - Completata
        * `FAILED` - Fallita
        * `CANCELLED` - Cancellata
security:
  - bearerAuth: []

# Esempi di utilizzo
...